backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
motor==3.7.0
pymongo==4.10.1
pydantic[email]==2.10.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
